加入安装 docker 和 node套件

+++ update_tools.d/mpfr_4.1.0.sh
@@ -8,14 +8,10 @@ mpfr_version="4.1.0"
 set -e
 
 # 设置代理
-source ./set_proxy.sh
+source ../set_proxy.sh
 
-if [ -f mpfr-${mpfr_version}.tar.xz ]; then
-    rm mpfr-${mpfr_version}.tar.xz
-fi
-if [ -d mpfr-${mpfr_version} ]; then
-    rm -r mpfr-${mpfr_version}
-fi
+[ -f mpfr-${mpfr_version}.tar.xz ] && rm mpfr-${mpfr_version}.tar.xz
+[ -d mpfr-${mpfr_version} ] && rm -r mpfr-${mpfr_version}
 
 # wget https://mirrors.tuna.tsinghua.edu.cn/gnu/mpfr/mpfr-4.1.0.tar.xz
 wget https://mirrors.tuna.tsinghua.edu.cn/gnu/mpfr/mpfr-${mpfr_version}.tar.xz
@@ -27,9 +23,5 @@ make -j 6
 sudo make install 
 cd ..
 
-if [ -f mpfr-${mpfr_version}.tar.xz ]; then
-    rm mpfr-${mpfr_version}.tar.xz
-fi
-if [ -d mpfr-${mpfr_version} ]; then
-    rm -r mpfr-${mpfr_version}
-fi
+[ -f mpfr-${mpfr_version}.tar.xz ] && rm mpfr-${mpfr_version}.tar.xz
+[ -d mpfr-${mpfr_version} ] && rm -r mpfr-${mpfr_version}
